test(mon-pix): factor out unabortedVisit
